:truck: Move Load/Unload Scene
